debundle/lowering: extract exports
BAZEL_TEST_INVOCATIONS=none: file move.
