+++ devinfra/js/debundle/BUILD.bazel
@@ -393,6 +393,7 @@ rust_library(
         "lowering/body_facts.rs",
         "lowering/chunk_ast.rs",
         "lowering/chunk_renames.rs",
+        "lowering/exports.rs",
         "lowering/imports_cross.rs",
         "lowering/imports_runtime.rs",
         "lowering/mod.rs",
